+++ internal/event/event.go
@@ -22,6 +22,7 @@ const (
 	TagKeyAccepted   = "halite/key/%s/accepted"
 	TagKeyRejected   = "halite/key/%s/rejected"
 	TagKeyRenewed    = "halite/key/%s/renewed"
+	TagKeyReissued   = "halite/key/%s/reissued"
 	TagJobDispatch   = "halite/job/%s/dispatch"
 	TagJobReturn     = "halite/job/%s/ret/%s"
 	TagBeacon        = "halite/beacon/%s/%s"